+++ tests/roots/test-docstrings/dummy_django_app/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+
+from . import views
+
+urlpatterns = [
+    path("simple/", views.simple_view),
+    path("simple/<int:year>/", views.simple_view),
+]